web_scraping_service/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
Flask-APScheduler>=1.13.1
pymongo>=4.6.1
Flask-PyMongo>=2.3.0
playwright>=1.40.0
weaviate-client>=4.4.b4